test: clear non-argument.type identifiers from PHPStan test baseline (PR 6/6) (#958)
* test: clear non-argument.type identifiers from PHPStan test baseline (PR 6/6)

Narrowed-assertion + style + judgment-tier sweep across ~190 test files:
alreadyNarrowedType family, property.notFound, ternary/cast/empty/equal
style tail, array.duplicateKey, method.internal, arrayFilter.strict, etc.
Terminal state: the only identifier remaining in phpstan-tests-baseline.neon
is the documented argument.type edge-case defer set (owned by #953).

Baseline note: the worktree container cannot run `analyse:tests` (symlinked-
vendor autoload yields false class.notFound; see docker_macos_autoload_realpath).
The test baseline here is the argument.type subset of post-#953 master
(77 entries / 152 occ) — CI's analyse:tests gate is the authoritative verifier.

Defects fixed beyond the nightly's automated sweep:
- ternary.shortNotAllowed expansion `f() ?: x` → `f() ? f() : x` double-evaluated
  the call (shell_exec git subprocess ×2, glob/scandir/class_implements/ini_get ×2
  across 8 files). Replaced with single-eval temp-var form, same predicate.
- cast.useless removals broke two DatabaseIntegration tests: mysqli returns
  COALESCE/SUM aggregates as numeric strings, so the removed (int)/(float) casts
  were load-bearing. BuyoutLedger: corrected the lying @var int → string and
  restored the casts (now meaningful → no cast.useless). ProjectedDraftOrder:
  assertSame → assertEquals (repo SUM type is string at runtime).
- RecordHolders: a removed meaningless-assertion else-branch left the null path
  assertion-free (risky on the seed); replaced with one path-independent assertion.

Verified: full non-DB suite green except pre-existing Tests\Cli\* env failures
(host bin/ not mounted in container); full --group database suite green
(871 tests, 5287 assertions).

* fix(test): restore assertSame with phpstan-ignore for SUM aggregate casts

assertEquals was introduced to avoid cast.useless, but triggered
phpunit.assertEquals instead. The cast is necessary at runtime (mysqli
returns SUM() as numeric-string); the @var float annotation overstates
the type. Annotate with @phpstan-ignore cast.useless per pattern from
feedback_cast_useless_db_aggregates.

* test: restore meaningful auth_user_id assertion in DevAutoLogin test

Code review flagged that the session user-id half of the "session
unchanged" contract was silently dropped. PHPStan had baselined the
assertion as staticMethod.alreadyNarrowedType because it narrows
$_SESSION['auth_user_id'] to the literal 42 from the setup assignment
and cannot see that tryAutoLogin() runs in between and could overwrite
it. The assertion is runtime-meaningful, so restore it with an inline
@phpstan-ignore (matching the SUM-aggregate restore pattern in this
branch) rather than re-adding a baseline entry.

+++ ibl5/tests/DatabaseIntegration/UpdateAllTheThings/PipelineIntegrationTestCase.php
@@ -174,7 +174,7 @@ abstract class PipelineIntegrationTestCase extends DatabaseTestCase
     protected function buildSchFile(array $games = []): string
     {
         $empty = str_repeat('0   0     ', SchFileParser::SLOTS_PER_DATE);
-        $data = str_repeat($empty, (int) (SchFileParser::FILE_SIZE / (SchFileParser::SLOTS_PER_DATE * SchFileParser::RECORD_SIZE)));
+        $data = str_repeat($empty, SchFileParser::FILE_SIZE / (SchFileParser::SLOTS_PER_DATE * SchFileParser::RECORD_SIZE));
 
         foreach ($games as $game) {
             $offset = ($game['date_slot'] * SchFileParser::SLOTS_PER_DATE + $game['game_index']) * SchFileParser::RECORD_SIZE;
